update PCRE to 8.21
git-svn-id: https://svn.r-project.org/R/trunk@58161 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/extra/pcre/pcre_valid_utf8.c
@@ -108,7 +108,7 @@ int _pcre_valid_utf8(USPTR string, int length, int *erroroffset)
     {
         for (p = string; *p != 0; p++)
             ;
-        length = p - string;
+        length = (int)(p - string);
     }
 
     for (p = string; length-- > 0; p++)
@@ -121,21 +121,21 @@ int _pcre_valid_utf8(USPTR string, int length, int *erroroffset)
 
         if (c < 0xc0) /* Isolated 10xx xxxx byte */
         {
-            *erroroffset = p - string;
+            *erroroffset = (int)(p - string);
             return PCRE_UTF8_ERR20;
         }
 
         if (c >= 0xfe) /* Invalid 0xfe or 0xff bytes */
         {
-            *erroroffset = p - string;
+            *erroroffset = (int)(p - string);
             return PCRE_UTF8_ERR21;
         }
 
         ab = _pcre_utf8_table4[c & 0x3f]; /* Number of additional bytes */
         if (length < ab)
         {
-            *erroroffset = p - string; /* Missing bytes */
-            return ab - length;        /* Codes ERR1 to ERR5 */
+            *erroroffset = (int)(p - string); /* Missing bytes */
+            return ab - length;               /* Codes ERR1 to ERR5 */
         }
         length -= ab; /* Length remaining */
 
@@ -143,7 +143,7 @@ int _pcre_valid_utf8(USPTR string, int length, int *erroroffset)
 
         if (((d = *(++p)) & 0xc0) != 0x80)
         {
-            *erroroffset = p - string - 1;
+            *erroroffset = (int)(p - string) - 1;
             return PCRE_UTF8_ERR6;
         }
 
@@ -159,7 +159,7 @@ int _pcre_valid_utf8(USPTR string, int length, int *erroroffset)
         case 1:
             if ((c & 0x3e) == 0)
             {
-                *erroroffset = p - string - 1;
+                *erroroffset = (int)(p - string) - 1;
                 return PCRE_UTF8_ERR15;
             }
             break;
@@ -171,17 +171,17 @@ int _pcre_valid_utf8(USPTR string, int length, int *erroroffset)
         case 2:
             if ((*(++p) & 0xc0) != 0x80) /* Third byte */
             {
-                *erroroffset = p - string - 2;
+                *erroroffset = (int)(p - string) - 2;
                 return PCRE_UTF8_ERR7;
             }
             if (c == 0xe0 && (d & 0x20) == 0)
             {
-                *erroroffset = p - string - 2;
+                *erroroffset = (int)(p - string) - 2;
                 return PCRE_UTF8_ERR16;
             }
             if (c == 0xed && d >= 0xa0)
             {
-                *erroroffset = p - string - 2;
+                *erroroffset = (int)(p - string) - 2;
                 return PCRE_UTF8_ERR14;
             }
             break;
@@ -193,22 +193,22 @@ int _pcre_valid_utf8(USPTR string, int length, int *erroroffset)
         case 3:
             if ((*(++p) & 0xc0) != 0x80) /* Third byte */
             {
-                *erroroffset = p - string - 2;
+                *erroroffset = (int)(p - string) - 2;
                 return PCRE_UTF8_ERR7;
             }
             if ((*(++p) & 0xc0) != 0x80) /* Fourth byte */
             {
-                *erroroffset = p - string - 3;
+                *erroroffset = (int)(p - string) - 3;
                 return PCRE_UTF8_ERR8;
             }
             if (c == 0xf0 && (d & 0x30) == 0)
             {
-                *erroroffset = p - string - 3;
+                *erroroffset = (int)(p - string) - 3;
                 return PCRE_UTF8_ERR17;
             }
             if (c > 0xf4 || (c == 0xf4 && d > 0x8f))
             {
-                *erroroffset = p - string - 3;
+                *erroroffset = (int)(p - string) - 3;
                 return PCRE_UTF8_ERR13;
             }
             break;
@@ -224,22 +224,22 @@ int _pcre_valid_utf8(USPTR string, int length, int *erroroffset)
         case 4:
             if ((*(++p) & 0xc0) != 0x80) /* Third byte */
             {
-                *erroroffset = p - string - 2;
+                *erroroffset = (int)(p - string) - 2;
                 return PCRE_UTF8_ERR7;
             }
             if ((*(++p) & 0xc0) != 0x80) /* Fourth byte */
             {
-                *erroroffset = p - string - 3;
+                *erroroffset = (int)(p - string) - 3;
                 return PCRE_UTF8_ERR8;
             }
             if ((*(++p) & 0xc0) != 0x80) /* Fifth byte */
             {
-                *erroroffset = p - string - 4;
+                *erroroffset = (int)(p - string) - 4;
                 return PCRE_UTF8_ERR9;
             }
             if (c == 0xf8 && (d & 0x38) == 0)
             {
-                *erroroffset = p - string - 4;
+                *erroroffset = (int)(p - string) - 4;
                 return PCRE_UTF8_ERR18;
             }
             break;
@@ -250,27 +250,27 @@ int _pcre_valid_utf8(USPTR string, int length, int *erroroffset)
         case 5:
             if ((*(++p) & 0xc0) != 0x80) /* Third byte */
             {
-                *erroroffset = p - string - 2;
+                *erroroffset = (int)(p - string) - 2;
                 return PCRE_UTF8_ERR7;
             }
             if ((*(++p) & 0xc0) != 0x80) /* Fourth byte */
             {
-                *erroroffset = p - string - 3;
+                *erroroffset = (int)(p - string) - 3;
                 return PCRE_UTF8_ERR8;
             }
             if ((*(++p) & 0xc0) != 0x80) /* Fifth byte */
             {
-                *erroroffset = p - string - 4;
+                *erroroffset = (int)(p - string) - 4;
                 return PCRE_UTF8_ERR9;
             }
             if ((*(++p) & 0xc0) != 0x80) /* Sixth byte */
             {
-                *erroroffset = p - string - 5;
+                *erroroffset = (int)(p - string) - 5;
                 return PCRE_UTF8_ERR10;
             }
             if (c == 0xfc && (d & 0x3c) == 0)
             {
-                *erroroffset = p - string - 5;
+                *erroroffset = (int)(p - string) - 5;
                 return PCRE_UTF8_ERR19;
             }
             break;
@@ -282,7 +282,7 @@ int _pcre_valid_utf8(USPTR string, int length, int *erroroffset)
 
         if (ab > 3)
         {
-            *erroroffset = p - string - ab;
+            *erroroffset = (int)(p - string) - ab;
             return (ab == 4) ? PCRE_UTF8_ERR11 : PCRE_UTF8_ERR12;
         }
     }
